requirements.txt
Flask==2.2.3
azure-storage-blob==12.19.0